spring boot 3.2 RestClient demo 추가

+++ src/main/java/com/example/declarativehttpinterface/demo/model/CurrencyApiResponse.java
@@ -1,10 +1,9 @@
 package com.example.declarativehttpinterface.demo.model;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 public interface CurrencyApiResponse {
-    Map<String, BigDecimal> getRates();
+    Map<String, Double> getRates();
     String getError();
     Boolean isSuccess();
 }
